feat(clay-button): create ClayButtonWithIcon component
