+++ packages/clay-button/package.json
@@ -26,6 +26,7 @@
 		"react"
 	],
 	"dependencies": {
+		"@clayui/icon": "^3.0.0-milestone.1",
 		"classnames": "^2.2.6"
 	},
 	"devDependencies": {
