+++ packages/clay-button/stories/index.tsx
@@ -4,10 +4,12 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 import '@clayui/css/lib/css/atlas.css';
-import ClayButton from '../src';
+import ClayButton, {ClayButtonWithIcon} from '../src';
 import React from 'react';
 import {boolean, select, text} from '@storybook/addon-knobs';
 import {storiesOf} from '@storybook/react';
+
+const spritemap = require('@clayui/css/lib/images/icons/icons.svg');
 
 storiesOf('ClayButton', module)
 	.add('default', () => (
@@ -42,4 +44,24 @@ storiesOf('ClayButton', module)
 			<ClayButton displayType="secondary">{'button'}</ClayButton>
 			<ClayButton>{'group.'}</ClayButton>
 		</ClayButton.Group>
+	))
+	.add('ButtonWithIcon', () => (
+		<>
+			<ClayButtonWithIcon spritemap={spritemap} symbol="trash" />
+			<ClayButtonWithIcon
+				displayType="secondary"
+				spritemap={spritemap}
+				symbol="cog"
+			/>
+			<ClayButtonWithIcon
+				displayType="link"
+				spritemap={spritemap}
+				symbol="cut"
+			/>
+			<ClayButtonWithIcon
+				displayType="unstyled"
+				spritemap={spritemap}
+				symbol="desktop"
+			/>
+		</>
 	));
